+++ isitfit/cost/cacheManager.py
@@ -85,7 +85,7 @@ And finally re-run isitfit as usual.
 """%n_ec2_total, "yellow"))
             continue_wo_redis = input(colored('Would you like to continue without redis caching (not recommended)? yes/[no] ', 'cyan'))
             if not (continue_wo_redis.lower() == 'yes' or continue_wo_redis.lower() == 'y'):
-                from isitfit.utils import IsitfitCliError
+                from isitfit.cli.click_descendents import IsitfitCliError
                 raise IsitfitCliError("Aborting to set up redis.", context_pre['click_ctx'])