+++ isitfit/cost/redshift_common.py
@@ -104,7 +104,7 @@ class CalculatorBaseRedshift:
     # Edit 2019-11-20 no need to through exception here
     # This way, the code can proceed to show a report, and possibly proceed to other services than redshift
     #if context_all['n_rc_analysed']==0:
-    #  from isitfit.utils import IsitfitCliError
+    #  from isitfit.cli.click_descendents import IsitfitCliError
     #  raise IsitfitCliError("No redshift clusters analyzed", context_all['click_ctx'])
 
     return context_all